Added more test code. Added 2 features. Deleted 1 feature.

+++ Sources/LHWWatcher.swift
@@ -26,12 +26,13 @@
 // THE SOFTWARE.
 
 import Foundation
+import CoreGraphics
 
 public protocol LHWWatcher: class {
     var actionHandler: LHWHandler? { get }
     
     func actorCompleted(status: LHWActionStageStatus, path: String, result: Any?)
-    func actorReportedProgress(path: String, progress: Float)
+    func actorReportedProgress(path: String, progress: CGFloat)
     func actionStageResourceDispatched(path: String, resource: Any?, arguments: Any?)
     func actionStageActionRequested(_ action: String, options: [String: Any]?)
     func actorMessageReceived(path: String, messageType: String?, message: Any?)
@@ -41,7 +42,7 @@ public extension LHWWatcher {
     func actorCompleted(status: LHWActionStageStatus, path: String, result: Any?) {
     }
     
-    func actorReportedProgress(path: String, progress: Float) {
+    func actorReportedProgress(path: String, progress: CGFloat) {
     }
     
     func actionStageResourceDispatched(path: String, resource: Any?, arguments: Any?) {
